modularize quickcache and isolate django dependency
