+++ quickcache/__init__.py
@@ -1,15 +1,16 @@
 from .quickcache import (
-    quickcache,
+    generic_quickcache,
     QuickCache,
-    TieredCache,
 )
+from .django_quickcache import quickcache
+
 from .deprecated import SkippableQuickCache, skippable_quickcache
 
 
 __all__ = [
+    'generic_quickcache',
     'quickcache',
     'skippable_quickcache',
     'QuickCache',
     'SkippableQuickCache',
-    'TieredCache',
 ]
